+++ keystore/base/base.go
@@ -1,62 +1,55 @@
-package keystore
+package base
 
 import (
 	"fmt"
 
 	ethcommon "github.com/ethereum/go-ethereum/common"
 	ethtypes "github.com/ethereum/go-ethereum/core/types"
+	"gitlab.com/ConsenSys/client/fr/core-stack/infra/key-store.git/keystore/session"
+	"gitlab.com/ConsenSys/client/fr/core-stack/infra/key-store.git/keystore/wallet"
 	"gitlab.com/ConsenSys/client/fr/core-stack/infra/key-store.git/secretstore"
 	"gitlab.com/ConsenSys/client/fr/core-stack/pkg.git/protos/common"
 )
 
-//BaseKeyStore olds the methods of the interfaces BaseKeyStore
-type BaseKeyStore struct {
+// KeyStore olds the methods of the interfaces BaseKeyStore
+type KeyStore struct {
 	SecretStore secretstore.SecretStore
 }
 
-//NewBaseKeyStore construct a BaseKeyStore from a client
-func NewBaseKeyStore(secretStore secretstore.SecretStore) *BaseKeyStore {
-
-	return &BaseKeyStore{
+//NewKeyStore construct a BaseKeyStore from a client
+func NewKeyStore(secretStore secretstore.SecretStore) *KeyStore {
+	return &KeyStore{
 		SecretStore: secretStore,
 	}
 }
 
 // SignTx returns a signed transaction. It is perfectly equivalent to SignTx
-func (s *BaseKeyStore) SignTx(
-	chain *common.Chain,
-	a ethcommon.Address,
-	tx *ethtypes.Transaction,
-) (raw []byte, hash *ethcommon.Hash, err error) {
-
-	sess := MakeTxSignature(s.SecretStore)
+func (s *KeyStore) SignTx(chain *common.Chain, a ethcommon.Address, tx *ethtypes.Transaction) (raw []byte, hash *ethcommon.Hash, err error) {
+	// Creates a new signing session
+	sess := session.MakeTxSignature(s.SecretStore)
 	err = sess.SetWallet(&a)
 	if err != nil {
 		return []byte{}, nil, err
 	}
-
 	sess.SetChain(chain)
 	sess.SetTx(tx)
 
+	// Run signing session
 	err = sess.Run()
 	if err != nil {
 		return []byte{}, nil, err
 	}
 
-	return sess.signedRaw, sess.txHash, nil
+	return sess.Raw, sess.Hash, nil
 }
 
 // SignMsg returns a signed message and its hash
-func (s *BaseKeyStore) SignMsg(
-	a ethcommon.Address,
-	msg string,
-) (rsv []byte, hash *ethcommon.Hash, err error) {
-
+func (s *KeyStore) SignMsg(a ethcommon.Address, msg string) (rsv []byte, hash *ethcommon.Hash, err error) {
 	return []byte{}, nil, fmt.Errorf("Not implemented yet")
 }
 
 // SignRawHash returns a signed raw hash
-func (s *BaseKeyStore) SignRawHash(
+func (s *KeyStore) SignRawHash(
 	a ethcommon.Address,
 	hash []byte,
 ) (rsv []byte, err error) {
@@ -65,9 +58,8 @@ func (s *BaseKeyStore) SignRawHash(
 }
 
 // GenerateWallet create and stores a new wallet in the vault
-func (s *BaseKeyStore) GenerateWallet() (add *ethcommon.Address, err error) {
-
-	wallet := NewWallet(s.SecretStore)
+func (s *KeyStore) GenerateWallet() (add *ethcommon.Address, err error) {
+	wallet := wallet.NewWallet(s.SecretStore)
 	err = wallet.Generate()
 	if err != nil {
 		return nil, err
@@ -78,13 +70,14 @@ func (s *BaseKeyStore) GenerateWallet() (add *ethcommon.Address, err error) {
 		return nil, err
 	}
 
-	return wallet.GetAddress(), nil
+	return wallet.Address(), nil
 }
 
 // ImportPrivateKey adds a private key in the vault
-func (s *BaseKeyStore) ImportPrivateKey(priv string) (err error) {
+// TODO: this is Unsafe and should be removed soon
+func (s *KeyStore) ImportPrivateKey(priv string) (err error) {
 
-	wallet := NewWallet(s.SecretStore)
+	wallet := wallet.NewWallet(s.SecretStore)
 	err = wallet.FromPrivateKey(priv)
 	if err != nil {
 		return err
@@ -95,17 +88,5 @@ func (s *BaseKeyStore) ImportPrivateKey(priv string) (err error) {
 		return err
 	}
 
-	return nil
-
-}
-
-// RegisterPkeys allow to register private keys on key store
-func (s *BaseKeyStore) RegisterPkeys(pkeys []string) (err error) {
-	for _, pkey := range pkeys {
-		err = s.ImportPrivateKey(pkey)
-		if err != nil {
-			return err
-		}
-	}
 	return nil
 }