+++ secretstore/hashicorp/config.go
@@ -1,12 +1,13 @@
-package secretstore
+package hashicorp
 
 import (
 	"fmt"
 
+	vault "github.com/hashicorp/vault/api"
+	log "github.com/sirupsen/logrus"
 	"github.com/spf13/pflag"
 	"github.com/spf13/viper"
-
-	vault "github.com/hashicorp/vault/api"
+	"gitlab.com/ConsenSys/client/fr/core-stack/infra/key-store.git/secretstore/aws"
 )
 
 func init() {
@@ -14,6 +15,10 @@ func init() {
 	viper.BindEnv(vaultURIViperKey, vaultURIEnv)
 	viper.SetDefault(vaultTokenNameViperKey, vaultTokenNameEnv)
 	viper.BindEnv(vaultTokenNameViperKey, vaultTokenNameEnv)
+	viper.SetDefault(vaultTokenViperKey, vaultTokenEnv)
+	viper.BindEnv(vaultTokenViperKey, vaultTokenEnv)
+	viper.SetDefault(vaultUnsealKeyViperKey, vaultUnsealKeyEnv)
+	viper.BindEnv(vaultUnsealKeyViperKey, vaultUnsealKeyEnv)
 }
 
 var (
@@ -28,7 +33,7 @@ var (
 	vaultTokenNameEnv      = "VAULT_TOKEN_NAME"
 
 	vaultTokenFlag     = "vault-token"
-	vaultTokenViperKey = "vault.token.value"
+	vaultTokenViperKey = "vault.token"
 	vaultTokenDefault  = ""
 	vaultTokenEnv      = "VAULT_TOKEN"
 
@@ -48,39 +53,34 @@ func InitFlags(f *pflag.FlagSet) {
 
 // VaultURI register a flag for vault server address
 func VaultURI(f *pflag.FlagSet) {
-	desc := fmt.Sprintf(`Hashicorp secret vault URI Environment variable: %q`, vaultURIEnv)
+	desc := fmt.Sprintf(`Hashicorp secret vault URI 
+Environment variable: %q`, vaultURIEnv)
 	f.String(vaultURIFlag, vaultURIDefault, desc)
-	viper.SetDefault(vaultURIViperKey, vaultURIDefault)
 	viper.BindPFlag(vaultURIViperKey, f.Lookup(vaultURIFlag))
-	viper.BindEnv(vaultURIViperKey, vaultURIEnv)
 }
 
 // VaultToken register a flag for vault server address
 func VaultToken(f *pflag.FlagSet) {
-	desc := fmt.Sprintf(`Hashicorp secret vault Token Environment variable: %q`, vaultTokenEnv)
+	desc := fmt.Sprintf(`Hashicorp secret vault authentication token
+Environment variable: %q`, vaultTokenEnv)
 	f.String(vaultTokenFlag, vaultTokenDefault, desc)
-	viper.SetDefault(vaultTokenViperKey, vaultTokenDefault)
 	viper.BindPFlag(vaultTokenViperKey, f.Lookup(vaultTokenFlag))
-	viper.BindEnv(vaultTokenViperKey, vaultTokenEnv)
 }
 
 // VaultUnsealKey registers a flag for the value of the vault unseal key
 func VaultUnsealKey(f *pflag.FlagSet) {
-	desc := fmt.Sprintf(`Hashicorp secret vault unseal key Environment variable: %q`, vaultTokenEnv)
+	desc := fmt.Sprintf(`Hashicorp vault unsealing key
+Environment variable: %q`, vaultTokenEnv)
 	f.String(vaultUnsealKeyFlag, vaultUnsealKeyDefault, desc)
-	viper.SetDefault(vaultUnsealKeyViperKey, vaultUnsealKeyDefault)
 	viper.BindPFlag(vaultUnsealKeyViperKey, f.Lookup(vaultUnsealKeyFlag))
-	viper.BindEnv(vaultUnsealKeyViperKey, vaultUnsealKeyEnv)
-
 }
 
 // VaultTokenName register a flag for vault server address
 func VaultTokenName(f *pflag.FlagSet) {
-	desc := fmt.Sprintf(`Hashicorp secret vault URI Environment variable: %q`, vaultTokenNameEnv)
+	desc := fmt.Sprintf(`Hashicorp token name on AWS
+Environment variable: %q`, vaultTokenNameEnv)
 	f.String(vaultTokenNameFlag, vaultTokenNameDefault, desc)
-	viper.SetDefault(vaultTokenNameViperKey, vaultTokenNameDefault)
 	viper.BindPFlag(vaultTokenNameViperKey, f.Lookup(vaultURIFlag))
-	viper.BindEnv(vaultTokenNameViperKey, vaultTokenNameEnv)
 }
 
 // NewConfig icreates vault configuration from viper
@@ -92,22 +92,50 @@ func NewConfig() *vault.Config {
 
 // AutoInit will try to Init the vault directly or FetchFromAws
 func AutoInit(hashicorps *Hashicorps) (err error) {
-	tokenName := viper.GetString("vault.token.name")
-	awsSS := NewAWS(7)
+	tokenName := viper.GetString(vaultTokenNameViperKey)
+	log.WithFields(log.Fields{
+		"aws.secret": tokenName,
+	}).Debugf("hashicorp: auto-initiliazing vault (credentials from AWS)")
+
+	// Create AWS object to retrieve Vault credentials
+	awsSS := aws.NewAWS(7)
+
+	// Initialize Vault
 	err = hashicorps.InitVault()
 	if err != nil {
+		log.WithError(err).Debugf("hashicorp: failed to init")
+
 		// Probably Vault is already unsealed so we retrieve credentials from AWS
-		err = hashicorps.InitFromAWS(awsSS, tokenName)
+		secret, err := awsSS.Load(tokenName)
 		if err != nil {
-			return fmt.Errorf("Could not retrieve credentials from AWS: %v", err)
-		}
-	} else {
-		// Vault has been properly unsealed so we push credentials on AWS
-		err = hashicorps.SendToCredStore(awsSS, tokenName)
-		if err != nil {
-			return fmt.Errorf("Could not send credentials to AWS: %v", err)
+			log.WithError(err).Errorf("hashicorp: failed to load credentials from AWS")
+			return err
 		}
 
+		err = hashicorps.creds.fromEncoded(secret)
+		if err != nil {
+			log.WithError(err).Errorf("hashicorp: failed to decode credentials")
+			return err
+		}
+
+		log.WithFields(log.Fields{
+			"vault.token": hashicorps.creds.Token,
+		}).Warnf("hashicorp: !!!WARNING this message is meant for debugging purpose and should be removed ASAP!!!")
+
+		// Set token
+		hashicorps.SetToken(hashicorps.creds.Token)
+	} else {
+		// Vault has been properly unsealed so we push credentials on AWS
+		encoded, err := hashicorps.creds.encode()
+		if err != nil {
+			return err
+		}
+
+		err = awsSS.Store(tokenName, encoded)
+		if err != nil {
+			log.WithError(err).Errorf("hashicorp: failed to store credentials in AWS")
+			return fmt.Errorf("Could not send credentials to AWS: %v", err)
+		}
 	}
 	return nil
 }